+++ examples/inputs-simple-sqlite/src/schema/__generated__/UserLast/index.ts
@@ -1,0 +1,5 @@
+export {
+  UserLastObject,
+  UserLastIdFieldObject,
+  UserLastNameFieldObject
+} from './object.base';