+++ src/utils/configUtils.ts
@@ -1,0 +1,6 @@
+import { ConfigInternal } from './config'
+
+export function getConfigCrudUnderscore(config: ConfigInternal) {
+  if (config.crud.underscoreBetweenObjectVariableNames) return '_'
+  return ''
+}